API can reset password #6071
This include a mailing system where the Message table is used as queue and
as a log of sent emails. Typically messages are queued instantly, and then
the Mailer is started by another process (cron) to send queued messages.
To send a message almost instantly we should use `sendMessageAsync()` which
will spawn a new process to be non-blocking.

For the password reseting part, a temporary token is generated when sending
the email and is the only thing needed to later actually change the password.

Also the CLI calls have the ACL filters completely disabled (in `cli.php`). This
allow cron scripts full access to everything. It's different from what we did
in the past, but for now it seems simpler and should fit our needs in this app.