Fix unit tests #6120
This also remove several `getTransactionLines()` in models because
calling repository from models should be avoided if possible. Instead
a proper Doctrine relation should be used. But even better is to not
do anything unless we are sure the client will need it. In most
cases the client should be able to use a relatively simple filter in
a separate graphql query to get the same result.

This avoid us work and make it more obvious to client that it might have
an impact on performance as it is separate SQL query.