Show preview of image to be uploaded even in successives uploads #6189
Previously only the first to-be-uploaded image was correctly previewed before its
upload. Then every successive upload on the same page, would actually show
the last already uploaded, and not the new to-be-uploaded image.

For the end-user it looked like the system "lagged" behind and was always late
of one image, which make it very confusing to use.