Delete correct file on disk when oprhaned image is deleted #6189
Doctrine seems to mixing up which entity is deleted when calling
lifcycle hooks in the context of a `orphanRemoval=true` thus leading
to us deleting the wrong file on disk and making out data incoherent.

This does **not** happen when the entity is properly loaded. While
I unfortunately don't understand why this is now thoroughly covered
by out unit tests.