AbstractModel.watchAll() support later subscription #6221
When calling watchAll() with existing variables, a GraphQL query will
immediately be send to server. Even if watchAll() user did not subscribe
to result. With this patch we can subscribe later and immediately receive
the answer that was received from server a while before.

This make watchAll much more robust and predictible without any downsides.