Keep the cart in sync between browser tabs #6394
Switching from sessionStorage to localStorage prevent a user to duplicate
orders by doing the following:

1. start an order
2. open a new tab which duplicate the order in the sessionStorage
3. go back to first tab
4. confirm the first order
5. come back on second tab
6. see that the (second) order is still visible
7. incorrectly think that the first order was not sent yet
8. click again to send the order, for the second time and thus duplicate it

localStorage allow use to share exactly the same cart across multiple
tabs in live and thus reduce confusion of what was sent or not.

The cart is cleared upon user login to avoid user1 starting an order, logging out.
And then user2 login in on the same browser and find an already started order.