Stop camera when cart is confirmed or emptied #6393
When a order is confirmed we can safely assume that the user won't be
scanning another item straight away, but he might not leave the page,
so we specifically stop the camera.

Similar when a cart is emptied. Maybe the whole shopping the user does
not want to buy anything anymore but "forget" to leave the page.