Edit stockMovement entire history #6410
This completely change the way stockMovement is managed.

First the product current quantity is always the same of all
stockMovements. So stockMovements cannot be deleted without impacting
the product quantity. If one day we have too many records in DB, we could
"compact" old stockMovements records into one per month or per year. But
the sum will have to stay identical.

Second product and stockMovement quantity are not managed by model anymore,
nor by DB triggers (because we cannot update the same table in a trigger). But
instead we use Doctrine lifecycle events. That means all stockMovement
manipulation MUST be done via the EntityManager. They MUST NOT be edited in
DB directly, and they MUST NOT be edited via SQL/DQL. Also they MUST NOT be
casacade deleted (unless if the whole product is being deleted).