Only the ErrorLink show error to end-user, never other components
Previously errors were shown differently whether a user passed through
the login page or not. This was because LoginComponent subscribed to
errors and failed to unsubscribe. By improving the ErrorLink to show
actual error message instead of a single generic hardcoded message, we
remove the need for special handling in LoginComponent simplifying the
whole thing.

We now show short generic message in case of internal errors, or a
long specific message as returned by server.