Favorize use of fragment for typing
By using uppercase for fragment it generates a more natural
type name that can then be used in our code to strongly type things
that comes from different queries in a simpler and more robust way.
Favorize use of fragment for typing

By using uppercase for fragment it generates a more natural
type name that can then be used in our code to strongly type things
that comes from different queries in a simpler and more robust way.