Always serve image format supported by browsers #7934
If uploading a TIFF, it must always be re-encoded into a format
that is supported by browsers. So either JPG or WebP.

In most cases, this is great because we will serve even more WebP
than before. Thus saving even more bandwidth.

But that means that if browser does not support WebP (Safari), asking
for an image at a specific size will always returns JPG and thus may
lead to the loss of transparency (if original is a format with
transparency such as PNG). AFAIK we don't have those use-cases in our
projects, and if transparency is a requirement then we must ask the image
without a specific size, in order to always get the original file (and
possibly receiving a poorly compressed file).