Prevent confirm-and-clear value with `enter` key #7966
The search bar in the hierarchic selector as a facet of natural-search
was a victim of confirm-and-clear bug. That means that when typing a
search string and confirming with ENTER, somehow the ENTER event was being
propagated to the button next to the input. And that button happened to be
the clear button. So that means that the value was confirmed (and searched),
but then immediately cleared (and search was cancelled). For humans it looked
impossible to use ENTER key to confirm a search.

It is not entirely clear why this behavior only appeared in a natural-search
in a hierarchic selector in a natural-search, and not in a simple natural-search.
But consuming the event seems to be a best-practice anyway.