Prevent infinite loop across browser tabs #8226
It might happen that we login in browser tab 1 and tab 2 get notified
so he will fetch viewer and update the local storage and thus notify
tab 1, who will do the same and enter an infinite loop.

This should not happen because the spec on https://html.spec.whatwg.org/multipage/webstorage.html#dom-storage-setitem
says that if the value has not changed (which is our case when tab 2 refresh his
viewer), then no event should be fired. However, despite the spec, it still happen
in some versions of Safari (see https://bugs.webkit.org/show_bug.cgi?id=211503)
and also in Chrome and Firefox as long as one tab is in an iframe (as seen in
Theodia various widgets).