New user must confirm before password reset #8244
If a user just created an account but did not confirmed yet, he can
request a password reset, but then instead of sending him the standard
reset email, we send him the confirm user email. So we force him to complete
his information and have a properly initialized user (with correct role),
before he can reset his password.

Otherwise it would let him logged in with the newly reset password with a
user that has incoherent role. And the application might look weird/broken
to him because of unexpected role.