Drop e2e tests #8336
Because they were never used and Protractor is now EOL. If
we need e2e tests in the future, we shall use Playwright instead.

Puppeteer is also replaced by Playwright for unit tests, because
it is easier to install system deps. And because it store browser
binaires outside of node_modules.