Typed GraphQL custom types #8180
Until now GraphQL custom types were all typed as `any` in Typescript. This
was overly permissive.

Now we can at least properly type some of them, making our whole code more robust.

But all "ID" types still cannot be typed properly, because we want to be able to
use either a string or an entire suboject. And typing that, and making it work
within Natural is difficult. So for now we still allow `any` as a compromise.