Use last value of resolvers #8992
Since Angular 14, resolvers will only take the first value. But our
resolvers typically emits multiple values, eg: cache then network. That
means that Angular 14 would show the cached version of an object on
its detail page. That becomes problematic if another user mutated that
object on the server after it was cached. It would never show the
mutated object.

To restore the behavior pre-Angular 14 we blindly added `last()` to all
resolvers, to be triple sure that we always use the value from network
and not cache. This addition may not be strictly necessary in all cases,
but better be safe than sorry for the time being. And we can always
remove the `last()` on a case by case basis after manual analysis.