MatTabGroup never stretch #9432
Tab stretching is a terrible UX where words float in the middle of
nowhere, making it extremely hard to understand their purpose in best
case scenario.

In the worst case scenario, then the second/third tab name will
accidentally align with the blocks contained in the first (selected) tab.
So the second/third tab name will actually look like they are the title
of the first tab content 🤮

Also, the fact that we lost the very fine line underlining the whole tab
group really makes the situation worse, but I guess we can at least live
that absence.