On upload most images are converted to WebP and have a max size of 3500px #9623
In production data we found many images that are very excessively too
large, such as some 68 MB JPG and/or that uses a very poor choice of
format, such as BMP, resulting in too big files.

This is a problem because dynamic generation of thumbnails will be
remarkably slow for the site visitor. And it is also a problem for
our storage that need to be over-dimensioned.