MariaDB 10.5.4 compatibility
Unfortunately, since https://jira.mariadb.org/browse/MDEV-22602, there
is a new rule that says we cannot use a FOREIGN KEY in a CHECK if that
FOREIGN KEY has a CASCADE rule. Because internally, MariaDB cannot
guarantee validate the CHECK when cascading happens. So from now on they
forbid the situation entirely.

This will be fixed but it might take a loooong time according to
https://jira.mariadb.org/browse/MDEV-25437?focusedCommentId=186537

In the meantime, we don't have any other choice but to drop the CHECK
entirely. Ideally, it should be restored when MariaDB supports again.