Flush pending mutation before resolving an object #10249
If we update an object, the mutation will be debounced for 2s. If we
quickly navigate to the list and click on the same object, then the
object will be re-fetched from server before the mutation is executed.
And we ultimately show a form that does not contain our latest changes.
Confusing the user that may then further overwrite other changes.

And always resolve object without cache at all