Disable passwords Safari "Smart quotes" for all password inputs #10443
If "Smart quotes" is enabled for Safari, that means that password
characters might get replaced by other characters without user noticing.
This means that password is not cross-browser compatible anymore. And a
password created in Safari cannot be used in any other browsers if it
contains characters subjected to "Smart quotes", or the other way around.

By unconditionally disabling spellcheck, we disable "Smart quotes", and
thus make password cross-browsers. However, users that created
their passwords on Safari with "Smart quotes" enabled (in the OS
settings),  will **no longer be able to login** if the password contains
characters such as `'`, `"`, `--` (and others ?).