Prevent not-submitting password value #10962
When app-password is initialized multiple times with the same form, then
`samePasswordsValidator` will already exist the second time we do
`removeControl`, and it will try to validate the form which now has a
missing control. Then JS will crash and leave the page in an undefined
state, which will allow submitting a mutation that does not specify any
password.

This is reproducible on grangette when failing the verification and
restarting the password update process.