Migrate to monolog and symfony/mailer #8635 #10257 #10724
Because laminas-log and laminas-mail are both discontinued. And we do it
now, because the migration to doctrine 3 exposed a cyclic dependency
issue in Epicerio. The end result was we were trying to log
something to DB before a connection to DB was established, and so we
established a second connection to DB, wrecking havoc in our apps.

So instead of deep diving in laminas-log to sort out all the details,
I'd rather deep dive in monolog directly and clear the technical debt at
the same time. And since monolog has an integration with symfony/mailer,
it seemed natural to do that too, and clear even more technical debt.

And to avoid the double-connection issue, the DB logger is only active,
and only retrieves its dependencies, after we are certain that a
connection to DB was established.

The overall concept of our logs remains unchanged though. We have 3 log
destinations: File, Mail and DB. Mail and DB have extra information
through `RecordCompleter`. File does not.