Prevent `doctrine_migration_versions` to be shown as non-sync error #11390
`./bin/doctrine orm:validate-schema` and `./bin/doctrine
orm:schema-tool:update --dump-sql` both incorrectly showed the DB
structure was out of sync because of the existence of the table
`doctrine_migration_versions`. The table must exist, and it must never
be deleted.

We reproduce the behavior of Doctrine older versions to ignore that
specific table for those two specific commands.

This also happens to create a base for a future Felix-dedicated CLI with
some commands defined within Felix and others within the application.